+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Add Scala 2.13 cross compilation [#310](https://github.com/azavea/stac4s/pull/310)
 
+### Changed
+- Make StacClient type alias a trait [#325](https://github.com/azavea/stac4s/pull/325)
+
 ## [0.4.0] - 2021-05-12
 ### Fixed
 - Item properties are no longer treated as a black box JsonObject [#309](https://github.com/azavea/stac4s/pull/309)