+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Add Scala 2.13 cross compilation [#310](https://github.com/azavea/stac4s/pull/310)
 - STAC Client pagination support [#327](https://github.com/azavea/stac4s/pull/327)
+- Brought compatibility up to 1.0.0 [#339](https://github.com/azavea/stac4s/pull/339)
 
 ### Changed
 - Make StacClient type alias a trait [#325](https://github.com/azavea/stac4s/pull/325)